+++ packages/@kagal-taistamp/eslint.config.mjs
@@ -2,5 +2,5 @@
 import { defineConfig } from '@poupe/eslint-config';
 
 export default defineConfig({
-  ignores: ['_docs', 'coverage'],
+  ignores: ['coverage'],
 });